standard transliteration from Russian
In Russian conversion into Latin, for several characters, several
transliterations are possible and used. Explicitly specify which ones
we want to use, and only use one value for each Cyrillic character.

This breaks round-trip transliteration from Polish-flavoured Cyrillic
(e.g. Jarosław > Ярослав > now Yaroslav), but the Polish-flavoured
is not a core feature, rather a bonus. The language supported
is specified as Russian, not more widely Cyrillic.